Move UserCard comonent out of Home.
This is due to it being reused in PersonalPage.
