+++ src/components/Home/Home.jsx
@@ -1,5 +1,5 @@
 import Search from "./Search/Search";
-import UserCard from "./UserCard/UserCard";
+import UserCard from "../UserCard/UserCard";
 import { useNavigate } from "react-router-dom";
 import { useState } from "react";
 import "./Home.css";
