+++ src/components/UserCard/TypeIcon/TypeIcon.css
