+++ src/components/UserCard/TypeIcon/TypeIcon.jsx
